+++ audio_modder.py
@@ -34,6 +34,7 @@ import config as cfg
 import db
 import log
 import fileutil
+import util
 
 from log import logger
 
@@ -87,21 +88,6 @@ CACHE = os.path.join(DIR, ".cache")
 # global variables
 language = 0
 num_segments = 0
-
-def language_lookup(lang_string):
-    try:
-        return LANGUAGE_MAPPING[lang_string]
-    except:
-        return int(lang_string)
-    
-def strip_patch_index(filename):
-    split = filename.split(".")
-    for n in range(len(split)):
-        if "patch_" in split[n]:
-            del split[n]
-            break
-    filename = ".".join(split)
-    return filename
     
 class WorkspaceEventHandler(FileSystemEventHandler):
 
@@ -188,185 +174,6 @@ class WorkspaceEventHandler(FileSystemEventHandler):
             elif str(child_path) == str(path):
                 return item
 
-def list_files_recursive(path="."):
-    files = []
-    if os.path.isfile(path):
-        return [path]
-    else:
-        for entry in os.listdir(path):
-            full_path = os.path.join(path, entry)
-            if os.path.isdir(full_path):
-                files.extend(list_files_recursive(full_path))
-            else:
-                files.append(full_path)
-        return files
-
-
-class MemoryStream:
-    '''
-    Modified from https://github.com/kboykboy2/io_scene_helldivers2 with permission from kboykboy
-    '''
-    def __init__(self, Data=b"", io_mode = "read"):
-        self.location = 0
-        self.data = bytearray(Data)
-        self.io_mode = io_mode
-        self.endian = "<"
-
-    def open(self, Data, io_mode = "read"): # Open Stream
-        self.data = bytearray(Data)
-        self.io_mode = io_mode
-
-    def set_read_mode(self):
-        self.io_mode = "read"
-
-    def set_write_mode(self):
-        self.io_mode = "write"
-
-    def is_reading(self):
-        return self.io_mode == "read"
-
-    def is_writing(self):
-        return self.io_mode == "write"
-
-    def seek(self, location): # Go To Position In Stream
-        self.location = location
-        if self.location > len(self.data):
-            missing_bytes = self.location - len(self.data)
-            self.data += bytearray(missing_bytes)
-
-    def tell(self): # Get Position In Stream
-        return self.location
-
-    def read(self, length=-1): # read Bytes From Stream
-        if length == -1:
-            length = len(self.data) - self.location
-        if self.location + length > len(self.data):
-            raise Exception("reading past end of stream")
-
-        newData = self.data[self.location:self.location+length]
-        self.location += length
-        return bytearray(newData)
-        
-    def advance(self, offset):
-        self.location += offset
-        if self.location < 0:
-            self.location = 0
-        if self.location > len(self.data):
-            missing_bytes = self.location - len(self.data)
-            self.data += bytearray(missing_bytes)
-
-    def write(self, bytes): # Write Bytes To Stream
-        length = len(bytes)
-        if self.location + length > len(self.data):
-            missing_bytes = (self.location + length) - len(self.data)
-            self.data += bytearray(missing_bytes)
-        self.data[self.location:self.location+length] = bytearray(bytes)
-        self.location += length
-
-    def read_format(self, format, size):
-        format = self.endian+format
-        return struct.unpack(format, self.read(size))[0]
-        
-    def bytes(self, value, size = -1):
-        if size == -1:
-            size = len(value)
-        if len(value) != size:
-            value = bytearray(size)
-
-        if self.is_reading():
-            return bytearray(self.read(size))
-        elif self.is_writing():
-            self.write(value)
-            return bytearray(value)
-        return value
-        
-    def int8_read(self):
-        return self.read_format('b', 1)
-
-    def uint8_read(self):
-        return self.read_format('B', 1)
-
-    def int16_read(self):
-        return self.read_format('h', 2)
-
-    def uint16_read(self):
-        return self.read_format('H', 2)
-
-    def int32_read(self):
-        return self.read_format('i', 4)
-
-    def uint32_read(self):
-        return self.read_format('I', 4)
-
-    def int64_read(self):
-        return self.read_format('q', 8)
-
-    def uint64_read(self):
-        return self.read_format('Q', 8)
-        
-def pad_to_16_byte_align(data):
-    b = bytearray(data)
-    l = len(b)
-    new_len = ceil(l/16)*16
-    return b + bytearray(new_len-l)
-    
-def _16_byte_align(addr):
-    return ceil(addr/16)*16
-    
-def bytes_to_long(bytes):
-    assert len(bytes) == 8
-    return sum((b << (k * 8) for k, b in enumerate(bytes)))
-
-def murmur64_hash(data, seed = 0):
-
-    m = 0xc6a4a7935bd1e995
-    r = 47
-
-    MASK = 2 ** 64 - 1
-
-    data_as_bytes = bytearray(data)
-
-    h = seed ^ ((m * len(data_as_bytes)) & MASK)
-
-    off = int(len(data_as_bytes)/8)*8
-    for ll in range(0, off, 8):
-        k = bytes_to_long(data_as_bytes[ll:ll + 8])
-        k = (k * m) & MASK
-        k = k ^ ((k >> r) & MASK)
-        k = (k * m) & MASK
-        h = (h ^ k)
-        h = (h * m) & MASK
-
-    l = len(data_as_bytes) & 7
-
-    if l >= 7:
-        h = (h ^ (data_as_bytes[off+6] << 48))
-
-    if l >= 6:
-        h = (h ^ (data_as_bytes[off+5] << 40))
-
-    if l >= 5:
-        h = (h ^ (data_as_bytes[off+4] << 32))
-
-    if l >= 4:
-        h = (h ^ (data_as_bytes[off+3] << 24))
-
-    if l >= 3:
-        h = (h ^ (data_as_bytes[off+2] << 16))
-
-    if l >= 2:
-        h = (h ^ (data_as_bytes[off+1] << 8))
-
-    if l >= 1:
-        h = (h ^ data_as_bytes[off])
-        h = (h * m) & MASK
-
-    h = h ^ ((h >> r) & MASK)
-    h = (h * m) & MASK
-    h = h ^ ((h >> r) & MASK)
-
-    return h
-
 class Subscriber:
     def __init__(self):
         pass
@@ -391,7 +198,6 @@ class AudioSource:
         self.data_OLD = b""
         self.subscribers = set()
         self.stream_type = 0
-        self.track_info = None
         
     def set_data(self, data, notify_subscribers=True, set_modified=True):
         if not self.modified and set_modified:
@@ -414,22 +220,7 @@ class AudioSource:
             
     def is_modified(self):
         return self.modified
-            
-    def set_track_info(self, track_info,  notify_subscribers=True, set_modified=True):
-        if not self.modified and set_modified:
-            self.track_info_old = self.track_info
-        self.track_info = track_info
-        if notify_subscribers:
-            for item in self.subscribers:
-                item.update(self)
-                if not self.modified:
-                    item.raise_modified()
-        if set_modified:
-            self.modified = True
-            
-    def get_track_info(self):
-        return self.track_info
-        
+
     def get_data(self):
         return self.data
         
@@ -458,7 +249,7 @@ class TocHeader:
     def __init__(self):
         pass
         
-    def from_memory_stream(self, stream):
+    def from_memory_stream(self, stream: MemoryStream):
         self.file_id             = stream.uint64_read()
         self.type_id             = stream.uint64_read()
         self.toc_data_offset     = stream.uint64_read()
@@ -494,7 +285,7 @@ class WwiseDep:
     def __init__(self):
         self.data = ""
         
-    def from_memory_stream(self, stream):
+    def from_memory_stream(self, stream: MemoryStream):
         self.offset = stream.tell()
         self.tag = stream.uint32_read()
         self.data_size = stream.uint32_read()
@@ -510,7 +301,7 @@ class DidxEntry:
         self.id = self.offset = self.size = 0
         
     @classmethod
-    def from_bytes(cls, bytes):
+    def from_bytes(cls, bytes: bytes | bytearray):
         e = DidxEntry()
         e.id, e.offset, e.size = struct.unpack("<III", bytes)
         return e
@@ -524,7 +315,7 @@ class MediaIndex:
         self.entries = {}
         self.data = {}
         
-    def load(self, didxChunk, dataChunk):
+    def load(self, didxChunk: bytes | bytearray, dataChunk: bytes | bytearray):
         for n in range(int(len(didxChunk)/12)):
             entry = DidxEntry.from_bytes(didxChunk[12*n : 12*(n+1)])
             self.entries[entry.id] = entry
@@ -540,7 +331,7 @@ class BankParser:
     def __init__(self):
         self.chunks = {}
         
-    def load(self, bank_data):
+    def load(self, bank_data: bytes | bytearray):
         self.chunks.clear()
         reader = MemoryStream()
         reader.write(bank_data)
@@ -554,7 +345,7 @@ class BankParser:
             size = reader.uint32_read()
             self.chunks[tag] = reader.read(size)
             
-    def GetChunk(self, chunk_tag):
+    def GetChunk(self, chunk_tag: str): -> bytearray
         try:
             return self.chunks[chunk_tag]
         except:
@@ -1673,7 +1464,7 @@ class FileHandler:
 
     def write_patch(self, folder=None):
         if folder == None:
-            folder = filedialog.askdirectory(title="Select folder to save files to")
+            raise Exception("No folder selected for patch generation")
         if os.path.exists(folder):
             patch_file_reader = FileReader()
             patch_file_reader.name = self.file_reader.name + ".patch_0"
@@ -1683,9 +1474,6 @@ class FileHandler:
             patch_file_reader.unknown = self.file_reader.unknown
             patch_file_reader.unk4Data = self.file_reader.unk4Data
             patch_file_reader.audio_sources = self.file_reader.audio_sources
-            patch_file_reader.string_entries = self.file_reader.string_entries
-            patch_file_reader.music_track_events = self.file_reader.music_track_events
-            patch_file_reader.music_segments = self.file_reader.music_segments
             patch_file_reader.wwise_banks = {}
             patch_file_reader.wwise_streams = {}
             patch_file_reader.text_banks = {}
@@ -1707,18 +1495,11 @@ class FileHandler:
             patch_file_reader.rebuild_headers()
             patch_file_reader.to_file(folder)
         else:
-            print("Invalid folder selected, aborting save")
-            return False
-        return True
+            raise Exception("Invalid folder selected for patch generation")
 
     def load_wems(self, wems: Union[tuple[str, ...], Literal[""], None] = None, set_duration=True): 
-        if wems == None:
-            wems = filedialog.askopenfilenames(title="Choose .wem files to import", filetypes=[("Wwise Vorbis", "*.wem")])
-        if wems == "":
-            return
-        progress_window = ProgressWindow(title="Loading Files", 
-                                         max_progress=len(wems))
-        progress_window.show()
+        if not wems:
+            raise Exception("No wems selected for import")
         for wem in wems:
             basename = os.path.basename(wem)
             splits: list[str] = basename.split("_", 1)
@@ -1728,11 +1509,10 @@ class FileHandler:
                         basename = name
             except:
                 pass
-            progress_window.set_text("Loading " + basename)
             file_id: int | None = self.get_number_prefix(basename)
             if file_id == None:
                 continue
-            audio: str | None = self.get_audio_by_id(file_id)
+            audio: AudioSource | None = self.get_audio_by_id(file_id)
             if audio == None:
                 continue
             with open(wem, 'rb') as f:
@@ -1766,10 +1546,8 @@ class FileHandler:
                             break
                     if stop:
                         break
-            progress_window.step()
-        progress_window.destroy()
-        
-    def create_external_sources_list(self, sources: list[str]):
+                        
+    def create_external_sources_list(self, sources: list[str]): -> str
         root = etree.Element("ExternalSourcesList", attrib={
             "SchemaVersion": "1",
             "Root": __file__
